Add LauncherRamp testing opmode and refactor ramp control
Introduces LauncherRampTesting opmode for testing launcher ramp functionality. Refactors Launcher and LauncherRamp classes for improved clarity, immutability, and documentation. Disables ColorSensorTesting opmode and adds my name to the top of a few of the files I wrote so that anyone who needs to know who to talk about problems with it will know.